requirements.txt
selenium==3.141.0
webdriver-manager==3.4.2
scikit-image==0.18.2
Pillow==8.3.1
numpy==1.21.2
opencv-python==4.5.3.56
tensorflow==2.5.0
